+++ src/app/beer_garden/events/processors.py
@@ -1,9 +1,14 @@
 # -*- coding: utf-8 -*-
 import logging
+import uuid
 from multiprocessing import Queue
 from queue import Empty
 
+from brewtils.models import Event, Events
 from brewtils.stoppable_thread import StoppableThread
+
+import beer_garden.config as config
+from beer_garden.queue.rabbit import put_event
 
 logger = logging.getLogger(__name__)
 
@@ -118,3 +123,47 @@ class FanoutProcessor(QueueListener):
 
         if manage:
             self._managed_processors.append(processor)
+
+
+class EventProcessor(FanoutProcessor):
+    """Class responsible for coordinating Event processing"""
+
+    def __init__(self, **kwargs):
+        super().__init__(**kwargs)
+        self.uuid = str(uuid.uuid4())
+
+    def put(self, event: Event, skip_checked: bool = False):
+        """Put a new item on the queue to be processed
+
+        Args:
+            event: New Event
+            skip_check: Flag to skip Event Name checks for routing
+        """
+
+        # Check if event should be published to Rabbit
+        if (
+            not skip_checked
+            and event.name != Events.REQUEST_TOPIC_PUBLISH.name
+            and (
+                event.name != Events.GARDEN_SYNC.name
+                or (
+                    event.name == Events.GARDEN_SYNC.name
+                    and event.garden != config.get("garden.name")
+                )
+            )
+        ):
+            try:
+                event.metadata["_source_uuid"] = self.uuid
+                put_event(event)
+                self._queue.put(event)
+            except Exception:
+                self.logger.error(f"Failed to publish Event: {event} to PIKA")
+                self._queue.put(event)
+        elif (
+            "_source_uuid" not in event.metadata
+            or event.metadata["_source_uuid"] != self.uuid
+        ):
+            self._queue.put(event)
+
+    def put_queue(self, event: Event):
+        self._queue.put(event)