+++ src/app/beer_garden/router.py
@@ -106,6 +106,7 @@ route_functions = {
     "REQUEST_READ": beer_garden.requests.get_request,
     "REQUEST_READ_ALL": beer_garden.requests.get_requests,
     "REQUEST_DELETE": beer_garden.requests.delete_requests,
+    "REQUEST_UPDATE": beer_garden.requests.update_request,
     "COMMAND_READ": beer_garden.commands.get_command,
     "COMMAND_READ_ALL": beer_garden.commands.get_commands,
     "INSTANCE_READ": beer_garden.systems.get_instance,
@@ -618,7 +619,7 @@ def _target_from_type(operation: Operation) -> str:
         else:
             return _instance_id_lookup(operation.args[0])
 
-    if operation.operation_type == "REQUEST_CREATE":
+    if operation.operation_type in ["REQUEST_CREATE", "REQUEST_UPDATE"]:
         target_system = System(
             namespace=operation.model.namespace,
             name=operation.model.system,
@@ -657,7 +658,30 @@ def _target_from_type(operation: Operation) -> str:
 
 def _system_name_lookup(system: Union[str, System]) -> str:
     with routing_lock:
-        return system_name_routes.get(str(system))
+        route = system_name_routes.get(str(system))
+
+    if route:
+        return route
+    else:
+        # If we don't know about the route, attempt to find it and update the routing table
+        systems = beer_garden.systems.get_systems(
+            namespace=system.namespace,
+            name=system.name,
+            version=system.version,
+        )
+        if len(systems) == 1:
+            for garden in get_gardens():
+                for system in garden.systems:
+                    if systems[0].id == system.id:
+                        with routing_lock:
+                            # Then add routes to systems
+                            add_routing_system(system=system, garden_name=garden.name)
+                        logger.error(
+                            "Router mapping is out of sync, you should consider re-syncing"
+                        )
+                        return garden.name
+
+    return None
 
 
 def _system_id_lookup(system_id: str) -> str: